+++ src/utils.ts
@@ -191,99 +191,3 @@ export const randHeader = () => {
   };
   return result;
 };
-
-export const fundRankHtmlTemp = (list: any[] = []) => {
-  let tbody = '';
-  const thead = `
-  <thead><tr ><th class="colorize">序号</th><th class="colorize">基金代码</th><th class="colorize">基金名称</th><th class="r_20 colorize">单位净值</th><th class="r_20 colorize">累计净值</th><th class="r_20">近三个月(%)</th><th class="r_20">近六个月(%)</th><th class=" r_20">近一年(%)</th><th class="sort_down r_20">今年以来(%)</th><th class=" r_20">成立以来(%)</th></tr></thead>`;
-  for (let i = 0; i < list.length; i++) {
-    const item = list[i];
-    const {
-      symbol,
-      name,
-      three_month,
-      six_month,
-      one_year,
-      form_year,
-      form_start,
-      dwjz,
-      ljjz,
-    } = item;
-    tbody += `<tr class="red">
-    <td class="colorize">${i + 1}</td>
-    <td class="colorize"><a href="http://biz.finance.sina.com.cn/suggest/lookup_n.php?q=${symbol}&amp;country=fund" target="_blank">${symbol}</a></td>
-    <td class="colorize"><a href="http://biz.finance.sina.com.cn/suggest/lookup_n.php?q=${symbol}&amp;country=fund" target="_blank" title="${name}" class="name">${name}</a></td>
-    <td class="r_20 colorize">${dwjz}</td>
-    <td class="r_20 colorize">${ljjz}</td>
-    <td class="r_20">${three_month}</td>
-    <td class="r_20">${six_month}</td>
-    <td class="r_20">${one_year}</td>
-    <td class="r_20 sort_down r_20">${form_year}</td>
-    <td class="r_20">${form_start}</td>
-    </tr>`;
-  }
-
-  return `<table boder="0">${thead}<tbody>${tbody} </tbody></table>`;
-};
-
-export const xuqiuArticleTemp = (newsList = [], hideAvatar = false) => {
-  const htmlArr = [];
-  for (let article of newsList) {
-    const info = article as any;
-    info.userId = info.user.id;
-    const images = info.user.profile_image_url.split(',');
-    const img = `https:${info.user.photo_domain}${images[images.length - 1]}`;
-    const description = info.description.replace(/\/\/assets/g, 'https://assets');
-
-    let articleStr = `
-    <article class="timeline__item">
-        ${
-          hideAvatar
-            ? ''
-            : `<a
-        href="https://xueqiu.com/${info.userId}"
-        target="_blank"
-        data-tooltip="${info.userId}"
-        class="avatar avatar-md"
-        ><img
-          src="${img}"
-      /></a>`
-        }
-        <div class="timeline__item__top__right"></div>
-        <div class="timeline__item__main" ${hideAvatar ? 'style="margin-left:0;"' : ''}>
-          <div class="timeline__item__info">
-            <div>
-              <a
-                href="https://xueqiu.com/${info.userId}"
-                target="_blank"
-                data-tooltip="${info.userId}"
-                class="user-name"
-                >${info.user.screen_name}</a
-              >
-            </div>
-            <a
-              href="https://xueqiu.com/${info.userId}/${info.id}"
-              target="_blank"
-              data-id="157971116"
-              class="date-and-source"
-              >${info.timeBefore} · 来自${info.source}</a
-            >
-          </div>
-          <div class="timeline__item__bd">
-            <div class="timeline__item__content">
-              <!---->
-              <div class="content content--description">
-                <!---->
-                <div class="">
-                  ${description}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </article>
-      `;
-    htmlArr.push(articleStr);
-  }
-  return htmlArr;
-};
